Extracting fetch logic
